Refactor: code to show user and edit user
