+++ client/package.json
@@ -19,6 +19,7 @@
     "react-loader-spinner": "^4.0.0",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
+    "react-toastify": "^8.0.3",
     "web-vitals": "^1.0.1"
   },
   "proxy": "http://localhost:5000",
