+++ client/src/App.js
@@ -1,4 +1,5 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
+import 'react-toastify/dist/ReactToastify.css';
 import { ProvideAuth } from './auth';
 import { AppRouter } from './routers';
