+++ client/src/components/inputs/DisableInput.js
@@ -4,8 +4,8 @@ export default function DisableInput({ disabled, ...props }) {
   return (
     <FormControl
       {...props}
-      readOnly={disabled}
       disabled={disabled}
+      value={props.value || ''}
       onChange={disabled ? null : props.onChange}
     />
   );
