+++ client/src/components/table/index.js
@@ -3,3 +3,4 @@ export * from './TableHead';
 export * from './TableBody';
 export * from './ActionsColumn';
 export { default as Pagination } from './Pagination';
+export { default as TableTitle } from './TableTitle';
